Integrate ingredients into the Pong game UI and create the victory screen

+++ pong/PongBall.py
@@ -4,7 +4,7 @@ from pong.PhysicsObject import PhysicsObject
 
 
 class PongBall:
-    def __init__(self, height: int, image_url: str, starting_location: Tuple[float, float], velocity: int):
+    def __init__(self, height: int, image_url: str, starting_location: Tuple[float, float], velocity: float):
         self.image_url = image_url
         self.physics_object = PhysicsObject(starting_location[0] - height // 2,
                                             starting_location[1] - height // 2,
